Add models from other projects [Regression] (#1718)
* finding missing workspace files

* using models from different projects enabled.

* cleaning up more.

* updated Versions.props
