+++ src/Shared/Microsoft.DotNet.Scaffolding.Shared/Microsoft.DotNet.Scaffolding.Shared.csproj
@@ -23,9 +23,8 @@
     <PackageReference Include="Microsoft.CodeAnalysis.VisualBasic.Features" Version="$(CodeAnalysisVersion)" />
     <PackageReference Include="Microsoft.CodeAnalysis.VisualBasic.Workspaces" Version="$(CodeAnalysisVersion)" />
     <PackageReference Include="Microsoft.CodeAnalysis.Workspaces.MSBuild" Version="$(CodeAnalysisVersion)" />
-    <PackageReference Include="Microsoft.Build" Version="$(MicrosoftBuildPackageVersion)" />
-    <PackageReference Include="Microsoft.Build.Utilities.Core" PrivateAssets="All" Version="$(MicrosoftBuildUtilitiesCorePackageVersion)" />
-    <PackageReference Include="Microsoft.Build.Tasks.Core" Version="$(MicrosoftBuildPackageVersion)" />
+    
+  
     <PackageReference Include="NuGet.ProjectModel" Version="$(NuGetProjectModelVersion)" />
   </ItemGroup>
   
@@ -34,12 +33,6 @@
       <DesignTime>True</DesignTime>
       <AutoGen>True</AutoGen>
       <DependentUpon>MessageStrings.resx</DependentUpon>
-    </Compile>
-    <Compile Include="$(RepoRoot)\src\Scaffolding\VS.Web.CG.Msbuild\ProjectContextWriter.cs">
-      <Link>Scaffolding\%(RecursiveDir)%(FileName)</Link>
-    </Compile>
-    <Compile Include="$(RepoRoot)\src\Scaffolding\VS.Web.CG.Msbuild\ProjectReferenceInformationProvider.cs">
-      <Link>Scaffolding\%(RecursiveDir)%(FileName)</Link>
     </Compile>
   </ItemGroup>
