+++ test/Scaffolding/VS.Web.CG.Core.Test/VS.Web.CG.Core.Tests.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <Import Project="$(RepoRoot)test\Scaffolding\TestPackage.props" />
